requirements.txt
fastapi 
uvicorn
openai
opencv-contrib-python
pillow
python-multipart
python-dotenv
supabase
mysqlclient 
sqlalchemy